seat: exposing default grab functions
