+++ compositor/input.c
@@ -29,7 +29,7 @@
 #include <xkbcommon/xkbcommon.h>
 
 #include <ctypes/helpers.h>
-#include <taiwins/objects/seat.h>
+#include <taiwins/objects/seat_grab.h>
 #include <taiwins/objects/logger.h>
 #include <taiwins/objects/utils.h>
