+++ libtaiwins/objects/data_device/data_dnd.c
@@ -26,7 +26,7 @@
 
 #include <taiwins/objects/utils.h>
 #include <taiwins/objects/data_device.h>
-#include <taiwins/objects/seat.h>
+#include <taiwins/objects/seat_grab.h>
 #include <taiwins/objects/surface.h>
 #include <taiwins/objects/cursor.h>
 #include <wayland-util.h>
